Moved smart searcher
