+++ forecasting_tools/agents_and_tools/base_rates/estimator.py
@@ -4,8 +4,8 @@ from typing import NamedTuple
 from forecasting_tools.agents_and_tools.deprecated.general_researcher import (
     GeneralResearcher,
 )
+from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
-from forecasting_tools.helpers.smart_searcher import SmartSearcher
 
 
 class EstimationResult(NamedTuple):
