+++ forecasting_tools/agents_and_tools/misc_tools.py
@@ -3,6 +3,7 @@ import asyncio
 from forecasting_tools.agents_and_tools.question_generators.simple_question import (
     SimpleQuestion,
 )
+from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
 from forecasting_tools.ai_models.agent_wrappers import AgentTool, agent_tool
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.ai_models.general_llm import GeneralLlm
@@ -12,7 +13,6 @@ from forecasting_tools.helpers.metaculus_api import (
     MetaculusApi,
     MetaculusQuestion,
 )
-from forecasting_tools.helpers.smart_searcher import SmartSearcher
 
 
 @agent_tool
