+++ forecasting_tools/forecast_bots/experiments/q1t_w_personas_and_exa.py
@@ -5,6 +5,7 @@ from datetime import datetime
 import typeguard
 from pydantic import BaseModel
 
+from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.ai_models.general_llm import GeneralLlm
 from forecasting_tools.data_models.forecast_report import ReasonedPrediction
@@ -23,7 +24,6 @@ from forecasting_tools.forecast_bots.official_bots.q1_template_bot import (
     Q1TemplateBot2025,
 )
 from forecasting_tools.helpers.prediction_extractor import PredictionExtractor
-from forecasting_tools.helpers.smart_searcher import SmartSearcher
 
 logger = logging.getLogger(__name__)
