+++ forecasting_tools/forecast_bots/experiments/q3t_w_exa.py
@@ -1,9 +1,9 @@
+from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.data_models.questions import MetaculusQuestion
 from forecasting_tools.forecast_bots.official_bots.q3_template_bot import (
     Q3TemplateBot2024,
 )
-from forecasting_tools.helpers.smart_searcher import SmartSearcher
 
 
 class Q3TemplateWithExa(Q3TemplateBot2024):
