+++ forecasting_tools/forecast_bots/official_bots/q1_template_bot.py
@@ -5,6 +5,7 @@ import os
 from datetime import datetime
 from typing import Literal
 
+from forecasting_tools.agents_and_tools.smart_searcher import SmartSearcher
 from forecasting_tools.ai_models.ai_utils.ai_misc import clean_indents
 from forecasting_tools.ai_models.general_llm import GeneralLlm
 from forecasting_tools.data_models.forecast_report import ReasonedPrediction
@@ -22,7 +23,6 @@ from forecasting_tools.forecast_bots.forecast_bot import ForecastBot
 from forecasting_tools.helpers.asknews_searcher import AskNewsSearcher
 from forecasting_tools.helpers.metaculus_api import MetaculusApi
 from forecasting_tools.helpers.prediction_extractor import PredictionExtractor
-from forecasting_tools.helpers.smart_searcher import SmartSearcher
 
 logger = logging.getLogger(__name__)
